requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.31
scikit-learn>=1.3.0
tensorflow>=2.13.0
newsapi-python>=0.2.7
textblob>=0.17.1
matplotlib>=3.7.0